04_assignments/assignments_01/Project 7 Weather Program Python Project/pyproject.toml
[project]
name = "project-7-weather-program-python-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
